Progress dump for full SDK flow
This looks like it's working now, which is great. It needs a lot of
tidying, splitting apart, commenting, testing, and error checking, but
the base functionality is there.

There's an odd change to `package.json` here. Because we're grabbing the
`package.json` file, it's shifting all of the files down a directory to
be relative to that import.