Fix function registered log spam
This fixes function registered log spam by hashing the contents of the
registration request, then storing the hash of the most recently
successful registration request.  If the hashes are the same
consecutively, we don't log a success message.

This also allows us to track hashes of each deploy from the devserver to
prevent log spam there :)