Test example functions and repos together on every PR (#37)
Introduces a new directory, `src/examples`, containing a collection of
example Inngest functions and step functions, e.g. `helloWorld.ts`,
`dripCampaign.ts`, etc.

With this, we then introduce a new PR check which will take in a list of
example repositories (e.g.
[inngest/sdk-example-cloudflare-workers](https://github.com/inngest/sdk-example-cloudflare-workers))
and for each:

- Clone both the SDK and the example repo
- Link the SDK locally to the example repo
- Copy the example functions from the SDK into the example repo's
`inngest` folder
- Assert that the example builds correctly with the SDK and the example
functions
- Boot the dev server for the example repo
- Assert basic functionality by curling the introspection endpoint

With this, we can add further assertions or tests now that we have a
place where the example repo is booted with example functions from the
SDK.

We could also venture to add test files for each example function, which
can attempt to access `http://localhost:3000/api/inngest` and assert
that that function is behaving correctly. I want to round off tests
today, though, and we may want to think about how the dev server fits
into this picture before going ham on integration tests.