fix: `require` not defined on landing page (#56)
**Problem**:
When trying to send event via landing page (For example:
http://localhost:8911/inngest), it throws error that `require` is not
defined.

![image](https://user-images.githubusercontent.com/2629902/207895642-ca2bd2e6-b794-4f07-bf05-05ca194a51e0.png)

**Fix:**
Avoid dynamic `require` import. We can use ES6 import but we are sharing
Inngest.ts for the npm module (commonjs) and landing page, so thought
this might be the safest way to go about it.

Let me know if you think of any changes. Also, feel free to discard this
PR if not needed.

cc: @KrisCoulson

Co-authored-by: Jack Williams <1736957+jpwilliams@users.noreply.github.com>