Refactor serve handlers (#52)
## Summary

Refactor handlers to be a strictly-typed, guided process instead of
having to extend a class and manually override specific methods.

Previously, all serve handlers would have to override a
`createHandler()` method on `InngestCommHandler` and ensure they handled
all eventualities (i.e. UI viewing, signing, registering, running,
etc.).

This was acceptable as we could lean on unit tests to ensure bases were
covered, but users writing custom handlers likely won't have this
luxury.

With this PR, each handler is now guided by strict types. The creator of
the handler specifies the input arguments and then defines how the SDK
will understand which action it should take. Once these are defined, the
creator is required to fulfil the correct payload for the action. For
example, running requires a function ID and data.

Finally, the user decides in one location how to transform the SDK's
response to an appropriate response for their framework.

This should be a much better experience for creating new handlers, as
well as providing a simple guide for users to create their own custom
handlers if they are needed.