Allow `sleepUntil` to parse strings
`tools.run` always JSON-serializes output, which means that if a user
returns a Date() then passes this into `tools.sleepUntil`, parsing
previously fails.

This brings up another issue:  `tools.run` typing auto-genned is
incorrect;  the resulting type is always serialized instead of being an
instance of whatever object you return.

For example, a `Set` is always an array, a `Date` is always a string, a
`Map` or `WeakMap`, is an object, classes are... tricky :)