fix: Ensure we return `200 OK` only when all operations are completed
If the function is returning a value, then it can't create any more
steps except by resolving or rejecting existing ones. If it does that,
we capture any next steps by stepping through the existing stack loop.

With this, we _can_ smartly identify when all operations in a function
have been completed via success or failure, which is awesome!

With this change, we keep returning `206 Partial Content` until every
Promise is settled, at which point that single, final request will
return `200 OK`.