Ensure `event|cron` union is strictly adhered to when declaring a func (#113)
## Summary

#110 was shipped in `v1.3.0` but reverted in
4c1ee8d388de9f0837b11db74a74134b8a9f7e66 due to causing an event typing
regression.

The fix is reintroduced here to resolve an issue where `event` and
`cron` could be passed in the same object when running
`inngest.createFunction()`. This includes some tests that came with the
revert in 4c1ee8d388de9f0837b11db74a74134b8a9f7e66 to ensure it's
working correctly.