Refactor how failure functions declare themselves to support serve opts
The function ID of the initial function is always required by the
failure function so that it can appropriately set an event expression.

IDs, however, are dynamic based upon the name and/or client given to a
specific serve handler. For this reason, the functions must remain
linked and produce a set of related configs instead of being two
entirely separate `InngestFunction` instantiations.

This also contains some extra smaller refactors around unifying some
constants and organizing code.