INN-1054 Ensure `ServeHandler` always returns `any` (#157)
## Summary INN-1054

Serve handlers used to all return any in the past, but have changed to
unknown in #135.

This is a mistake. We need any so that the output can be appropriately
slotted into any requirement the user has.

`any` feels dangerous here, but is appropriate: we can infer the types
for the returned function appropriately, e.g. `(req: NextRequest, res:
NextResponse) => NextResponse`, but this is also a misstep. To do this,
we must install a package for the `NextRequest` and `NextResponse`
types, but the package we use could be different to the one the user
uses. This can create a drift in types and cause a collision where there
shouldn't be one.

The fix is just to revert all handlers to appropriately return `any`.

## Related

- #135
- INN-1054