INN-1126 Execute a step early if it's the only pending item during a discovery (#165)
## Summary INN-1126

Steps within a function are always "planned" before being executed. This
means there are always two executions instead of one, even for very
simple functions: one that reports the step and one that executes the
step.

This provides us with a couple of benefits:

1. **Safety** - By ensuring that multiple steps aren't executed
simultaneously, we ensure that every execution is scoped to a particular
step; if the execution fails, we know exactly which step has failed.

2. **Configuration changes** - Some pieces of configuration, such as
setting how many retries a particular step has, must be reported to the
executor before execution, as otherwise, this information may not be
correctly presented.

We can still provide a short-circuit here, ensuring that steps are
executed immediately if they'll be the only reported step.