No longer manage `loggerMiddleware` inside `logger.ts` :(
Some TS comment inference annoyance means we have to define this inside
the client instead. It's no impact for the user, but it's annoying to
have to do.

If we can figure out a nice way to infer these tuples along with their
comments then we can move it back!