Ensure `getHookStack()` requires transforms for functions with skewed IO
While we basically do a reduce over middleware, we want to provide the
writer of middleware with a prettier interface, ensuring they don't have
to return the entire next argument. This is _also_ required for
appropriately typing mutated arguments, ensuring we're not accidentally
overwriting existing arguments with more generic ones from middleware.

So, the input to a middleware mutation function like `run->input` is
different from the expected output. For this reason, we enforce that any
hooks with a differing input and output must provide a typed `transform`
function which will take output from one function in the chain and use
that to prepare the input to the next.