Add workflow that allows us to automatically backport merged PRs (#223)
## Summary

This handy action allows us to backport a PR to a particular branch when
merged using labels.

![image](https://github.com/inngest/inngest-js/assets/1736957/54f4295d-8bc8-4341-a277-0344f9e86615)

Just add the label, then when the PR is merged a new PR will be opened
to the relevant branch (e.g. `1.x`) to backport the squashed commit.

This does not manage releasing; we'll use changesets for that.

## Related

- #224

(cherry picked from commit efc37dcf5446a6c7f0c52b5fe7d410729518ae55)