Include 'modified' in register response (#250)
# Changes

Include `modified` in register response.

# Purpose

The API will be able to forward the `modified` value to the UI during
registration. This will give the UI enough information to know whether a
deploy was created or deduped. We can use that to improve the deploy
toast notification in the UI, like saying "No deploy created because
function configuration is unchanged".

# Testing

I manually tested this with an API feature branch and the `modified`
value in the response is `true` if a deploy was created and `false` if
the deploy was deduped.

---------

Co-authored-by: Jack Williams <jack@inngest.com>