chore: update cross-fetch to ^4.0.0 package (#312)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

This updates the `cross-fetch` package to `^4.0.0` which drops support
for node 10 and 12 (this package uses node >=14) + adds official support
for the current LTS node 18 and node 20. `cross-fetch` also updated some
implementations around the Cloudflare worker fetch implementation as has
a more solid detection over supported fetch implementations.

Secondly I updated the pnpm package manager to 8.6.2 which added a fix
for the lockfile which made some troubles

https://github.com/pnpm/pnpm/issues/6648#issuecomment-1588863457 


## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~
- [ ] ~~Added unit/integration tests~~ Already exist
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->