Add ability to declare and send events without payloads (#313)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Add the ability to declare and send events without payloads. It's
possible for some events to rightfully have no payload, in which case
just the name can suffice.

```ts
// Previously would have failed
inngest.send({ name: "my.event" });

// Typing a no-data event means providing an empty object
new EventSchemas()
  .fromRecord<{ "my.event": {} }>()
  .fromUnion<{ name: "my.event" }>()
  .fromZod({ "my.event": {} })
  .fromZod([z.object({ name: z.literal("my.event") })]);
```

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A No need to directly recommend this
- [x] Added unit/integration tests
- [x] Added changesets if applicable