Simplify step returns such that `JsonifyObject<>` is hidden entirely (#371)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Step results are serialized and deserialized while being pushed to/from
Inngest, so step returns are typed as JSON to make this clear.

These returns still came back wrapped in `JsonifyObject<>`, though,
meaning they didn't actually show the resulting changes unless the user
started to attempt to use them. This small change ensures we display the
resulting type correctly, free of any wrappers.

|
![image](https://github.com/inngest/inngest-js/assets/1736957/cc1027a3-d246-43e1-b6c9-7adad2dea485)
|
![image](https://github.com/inngest/inngest-js/assets/1736957/1c639a86-cd2f-49a1-96cd-2018c7ba52de)
|
| - | - |
| Before | After |


## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A
- [ ] ~~Added unit/integration tests~~ N/A
- [x] Added changesets if applicable