Fix `stringify()` failing to serialize `BigInt`; remove instead (#375)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->


[`BigInt`](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/BigInt)s
are not natively parsed by JSON. Attempting to return a `BigInt` from a
step or function will throw an error. This error was hard to see, which
is fixed in #374.

Our types reflect that `BigInt` is unparseable and will type it as
`never`, though these were obscured from the user a touch, which is
fixed in #371.

For this reason, this PR ensures `BigInt`s are removed during parsing
instead of throwing. We can investigate adding `BigInt` support across
our SDKs past this fix; the `n` suffix seems an unwritten standard.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- #371
- #374