Add CI tests for dist and dependency issues (#406)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

An assortment of fixes and regression tests for consumers using
`skipLibCheck: false`.

New CI test checks for dependencies used within exported files but in
`devDependencies` instead of `dependencies`. For any regular project,
all types would be inside `devDependencies`, but this isn't true for
libraries depending on how they're being consumed.

We use `import/no-extraneous-dependencies` for most of this (thanks,
@goodoldneon), but include the script to also cover `@types/*` packages
that must be shipped in `dependencies` or bundled. This can be the case
if the consuming package won't get types from the base package that _is_
included in `dependencies` and those types are part of any exported
types.

> [!NOTE]
> We currently ignore modules within serve handler files, as we expect
these types to be peer dependencies.

Successful builds can also still produce erroring `*.d.ts` files against
some versions/configurations of TypeScript. Another new CI test runs
checks against these generated files with every supported minor of
TypeScript.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A Avoidance of this error no longer needed
- [x] Added unit/integration tests
- [x] Added changesets if applicable