Fix Sentry/Express tracing issues when using `app.use()` (#440)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Express performs some runtime checks to assess functions being used to
handle middleware. When used with Express Sentry middleware (which wraps
other endpoints, like Inngest's), our internal use of variadic arguments
can produce a function signature that looks like it can't adequately
handle a request.

This should be resolved by using the signature of the function passed
when creating the serve handler, ensuring we keep the function signature
looking healthy.

Thanks for the repro of this issue, @spastorelli! Looks fixed using this
PR (`inngest@pr-440`).

- Original repro:
https://replit.com/@spsgitbook/InngestWithSentryRepro#index.js
- Fixed repro:
https://replit.com/@jpwilliams1/InngestWithSentryRepro#index.js
`npm install inngest@pr-440`

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A Bug fix
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- Fixes #436
- getsentry/sentry-javascript#3284