Fix a TS error when using Inngest in projects with `composite` setting (#460)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->
This PR Adds some type exports to `index.ts` that fix a TS error in
composite projects.

After changes to the `types.ts` the issue #384 appeared again. This PR
is similar to #385 - it fixes the TS error by adding necessary exports
to the `index.ts`

## Checklist


- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A no docs needed
- [ ] ~~Added unit/integration tests~~ N/A not testable
- [x] Added changesets if applicable

## Related
- #385 
- #384

---------

Co-authored-by: Jack Williams <1736957+jpwilliams@users.noreply.github.com>