Add experimental Bun example (#437)
## Summary

An example using [Bun](https://bun.sh/) and the existing `edge` serve
handler.

### Notes

* There is a module resolution issue with the default `tsconfig.json`
generated via `bun init` which can be "fixed" by changing the
`"moduleResolution"` to `"node"`. `The inferred type of 'helloWorld'
cannot be named without a reference to '../node_modules/inngest/types'.
This is likely not portable. A type annotation is necessary.`

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [x] Added a [docs PR](https://github.com/inngest/website) that
references this PR
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related

- inngest/website#666

---------

Co-authored-by: Jack Williams <1736957+jpwilliams@users.noreply.github.com>