Use `StepError`s for unhandled error catching (#474)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Aligns with the
[spec](https://github.com/inngest/inngest/blob/main/docs/SDK_SPEC.md)
for throwing errors when memoizing data. This previous solution was for
the initial release of `step.invoke()` pre-per-step errors.

Functionally there is no change, though the UI will no longer show
`NonRetriableError` prefixes and will prefer user error data.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A
- [x] Added unit/integration tests
- [x] Added changesets if applicable