Fix delayed environment variables not using event key (#492)
## Summary
<!-- Succinctly describe your change, providing context, what you've
changed, and why. -->

Fixes an issue where delayed access to environment variables was
incorrectly missing setting the event key.

`Inngest.eventKey` will always be a populated string, falling back to a
dummy value of `NO_EVENT_KEY_SET` if it's not specified during
instantiation. We do this as API routing for ingestion
(`https://inn.gs/e/:event_key`) sensibly relies on the event key being
present, even in local dev.

In environments such as Cloudflare Workers where access to environment
variables is delayed until the request handler is called, we'll should
set the event key again if we only have a dummy and we now have a real
one.

This code was erroneously checking that `eventKey` was truthy instead of
calling the existing `eventKeySet()` method that considers dummy keys.

## Checklist
<!-- Tick these items off as you progress. -->
<!-- If an item isn't applicable, ideally please strikeout the item by
wrapping it in "~~"" and suffix it with "N/A My reason for skipping
this." -->
<!-- e.g. "- [ ] ~~Added tests~~ N/A Only touches docs" -->

- [ ] ~~Added a [docs PR](https://github.com/inngest/website) that
references this PR~~ N/A Bug fix
- [x] Added unit/integration tests
- [x] Added changesets if applicable

## Related
<!-- A space for any related links, issues, or PRs. -->
<!-- Linear issues are autolinked. -->
<!-- e.g. - INN-123 -->
<!-- GitHub issues/PRs can be linked using shorthand. -->
<!-- e.g. "- inngest/inngest#123" -->
<!-- Feel free to remove this section if there are no applicable related
links.-->
- Fixes #481